Consolidate Rust learning into single 'Rust By Example' folder
- Move polished course notes (L1-L4) into Rust By Example/src/bin/
  so each lesson file serves as both runnable demo and study notes.
- Add README.md with project layout, conventions, and curriculum TOC.
- Remove now-redundant 'Rust Notes/' folder.
- Remove src/bin/003-shadowing.rs (intentionally-broken exercise;
  shadowing is fully covered in the polished 002-mutability.rs).

src/main.rs is preserved as the active-lesson scratchpad.
