+++ Rust By Example/src/bin/003-data-types.rs
