--- Rust By Example/src/bin/003-shadowing.rs
@@ -1,3 +1,0 @@
-fn main() {
-
-}
