+++ Rust By Example/src/bin/004-functions.rs
