--- Rust Notes/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "rust-notes"
-version = "0.1.0"
-edition = "2024"
-default-run = "rust-notes"
-
-[dependencies]
